Added Padding

+++ src/VegasTools/Models/KeyFrameInfo.cs
@@ -2,10 +2,10 @@ namespace VegasTools.Models
 {
     public class KeyFrameInfo
     {
-        public double time;
-        public double PanX;
-        public double PanY;
-        public double Width;
-        public double Height;
+        public decimal time;
+        public decimal PanX;
+        public decimal PanY;
+        public decimal Width;
+        public decimal Height;
     }
 }
